Implemented new exception handler.
A new exception handler will handle any uncaught exceptions, so only a short
error message gets printed on stderr. In future developement this makes the code
way more simple, as more exceptions may be raised instead of catching them at
any time in the code to print beautiful messages. In addition that makes
debugging easier, as the traceback may be enabled on-demand.