Added job-step class for future use.
A new class will manage the job steps at a global- and job-level and will be
used in updated classes for pipelines and jobs in the future. The class converts
all commands into lists and gives compressed output on dump, so the handling
is very easy and simplifies the handling inside the dispatcher and runner.