JobSteps class support parent namespaces now.
Pipelines as well as jobs may define job steps - whenever a job doesn't have a
specialized step, the one of the pipeline will be used, if available. The
JobStep class is now able to handle a hierarchy: Whenever itself doesn't have
commands for a specific step, it will lookup the step in a parent namespace.