Moved pipeline creation into dedicated class.
A new child-class will handle the creation of a new pipeline, so the methods
required for creating a new pipeline will only exist in this namespace.

This is a first step for separating read- and write-access to a pipeline, as the
runner requires locking of pipelines for concurrent access. This can only be
achieved by limiting the access outside of the critical region to force the use
of standardized methods for changing the pipeline's data safely.