Pipeline handles concurrent access now.
The James CI runner may be run in parallel for several jobs. To prevent data
corruption in the pipeline's configuration file, this file will be locked as in
the legacy pipeline class. Concurrent processes may read the configuration file,
but just a single process will have access to this file at the same time for
writing into it.